+++ site/src/other.rs
@@ -504,6 +504,56 @@ pub fn Subscripts() -> impl IntoView {
         </tr>)
     }
 
+    let stable = vec![
+        subscript(Couple, "Group N arrays as rows", "{⊟₃ 1 2 3 4 5}"),
+        subscript(Box, "Group N arrays as boxed rows", "□₃ 1_2_3 5 \"wow\""),
+        subscript(Deshape, "Change rank", "♭₂ ⇡ 2_2"),
+        subscript(Transpose, "Repeat", "# Experimental!\n△ ⍉₃ °△1_2_3_4_5"),
+        subscript(Sqrt, "Nth root", "√₃ [8 27 125]"),
+        subscript(Round, "To N decimal places", "⁅₃ π"),
+        subscript(Floor, "To N decimal places", "# Experimental!\n⌊₄ π\n⌊₄ τ"),
+        subscript(Ceil, "To N decimal places", "# Experimental!\n⌈₄ π\n⌈₄ τ"),
+        subscript(First, "First N values", "⊢₂ \"hello\""),
+        subscript(Last, "Last N values", "⊣₂ \"hello\""),
+        subscript(Rand, "Random integer", "# Experimental!\n⚂₁₀₀"),
+        subscript(Both, "Apply to N argument sets", "[∩₃+ 1 2 3 4 5 6]"),
+        subscript(Each, "Apply to rank N subarrays", "∵₁□ °△2_3_4"),
+        subscript(Rows, "Apply to subarrays N deep", "≡₂□ °△2_3_4"),
+        subscript(Inventory, "Apply to subarrays N deep", "⍚₂⇡ °△2_3"),
+        subscript(Repeat, "Repetition count", "⍥₅(⊂⟜/+) [1 2]"),
+        subscript(Tuples, "Tuple size", "# Experimental!\n⧅₂< ⇡4"),
+        subscript(Stencil, "Window size", "# Experimental!\n⧈₃∘ ⇡6"),
+        subscript(Stack, "Print top N values", "?₂ 1 2 3 4"),
+    ];
+
+    let sided = vec![
+        subscript(
+            Both,
+            "Use left-most or right-most argument twice",
+            "# Experimental!\n[∩⌞⊟ @a@b@c]\n[∩⌟⊟ @a@b@c]",
+        ),
+        subscript(
+            Bracket,
+            "Use left-most or right-most argument twice",
+            "# Experimental!\n[∩⌞⊟ @a@b@c]\n[∩⌟⊟ @a@b@c]",
+        ),
+        subscript(
+            Rows,
+            "Fix left-most or right-most argument",
+            "# Experimental!\n≡⌞⊂ 1_2_3 4_5_6\n≡⌟⊂ 1_2_3 4_5_6",
+        ),
+        subscript(
+            Inventory,
+            "Fix left-most or right-most argument",
+            "# Experimental!\n⍚⌞⊂ 1_2_3 4_5_6\n⍚⌟⊂ 1_2_3 4_5_6",
+        ),
+        subscript(
+            Each,
+            "Fix left-most or right-most argument",
+            "# Experimental!\n∵⌞⊂ 1_2 3_4\n∵⌟⊂ 1_2 3_4",
+        ),
+    ];
+
     view! {
         <Title text="Subscripts - Uiua Docs"/>
         <h1>"Subscripts"</h1>
@@ -526,27 +576,24 @@ pub fn Subscripts() -> impl IntoView {
                 <td>"Constant first argument"</td>
                 <td><Editor example="⊃↙₂↻₃ [1 2 3 4 5]" nonprogressive=true/></td>
             </tr>
-            { vec![
-                subscript(Couple, "Group N arrays as rows", "{⊟₃ 1 2 3 4 5}"),
-                subscript(Box, "Group N arrays as boxed rows", "□₃ 1_2_3 5 \"wow\""),
-                subscript(Deshape, "Change rank", "♭₂ ⇡ 2_2"),
-                subscript(Transpose, "Repeat", "# Experimental!\n△ ⍉₃ °△1_2_3_4_5"),
-                subscript(Sqrt, "Nth root", "√₃ [8 27 125]"),
-                subscript(Round, "To N decimal places", "⁅₃ π"),
-                subscript(Floor, "To N decimal places", "# Experimental!\n⌊₄ π\n⌊₄ τ"),
-                subscript(Ceil, "To N decimal places", "# Experimental!\n⌈₄ π\n⌈₄ τ"),
-                subscript(First, "First N values", "⊢₂ \"hello\""),
-                subscript(Last, "Last N values", "⊣₂ \"hello\""),
-                subscript(Rand, "Random integer", "# Experimental!\n⚂₁₀₀"),
-                subscript(Both, "Apply to N argument sets", "[∩₃+ 1 2 3 4 5 6]"),
-                subscript(Each, "Apply to rank N subarrays", "∵₁□ °△2_3_4"),
-                subscript(Rows, "Apply to subarrays N deep", "≡₂□ °△2_3_4"),
-                subscript(Inventory, "Apply to subarrays N deep", "⍚₂⇡ °△2_3"),
-                subscript(Repeat, "Repetition count", "⍥₅(⊂⟜/+) [1 2]"),
-                subscript(Tuples, "Tuple size", "# Experimental!\n⧅₂< ⇡4"),
-                subscript(Stencil, "Window size", "# Experimental!\n⧈₃∘ ⇡6"),
-                subscript(Stack, "Print top N values", "?₂ 1 2 3 4"),
-            ] }
+            { stable }
+        </table>
+        <Hd id="sided">"Sided Subscripts"</Hd>
+        <p>"Sided subscripts are an "<code>"# Experimental!"</code>" feature that allows using subscripts for some common patterns that might be thought of as having a \"side\"."</p>
+        <p>"Sided subscripts are typed like normal subscripts with "<code>"__"</code>", but followed by "<code>"<"</code>" for left or "<code>">"</code>" for right. The formatter will turn them into "<code>"⌞"</code>" and "<code>"⌟"</code>" respectively."</p>
+        <p>"Sided and numberic subscripts cannot currently be mixed."</p>
+        <table class="header-centered-table cell-centered-table" style="width: 100%">
+            <tr>
+                <th>"Primitive"</th>
+                <th>"Meaning"</th>
+                <th style="width: 60%">"Example"</th>
+            </tr>
+            <tr>
+                <td>"Any "<span class="dyadic-function">"dyadic"</span>" pervasive function"</td>
+                <td>"Fix left-most or right-most argument"</td>
+                <td><Editor example="# Experimental!\n+⌞ ×100. 1_2_3\n+⌟ ×100. 1_2_3" nonprogressive=true/></td>
+            </tr>
+            { sided }
         </table>
     }
 }